Task 2: Modify the Checkout Process - Added gift message - Checkout/Confirm screen - Added gift message field, made required change in opc js - Added CustomOrderDefaults - Added customResource xml - Admin - Customized order module to show gift message - Public/OrderDetail - Customized orderModelFactory, page and model to show gift message - Added customized view in theme
Note: Custom code in .cs file is added with ++Custom code comment
