+++ src/Libraries/Nop.Core/CustomCode/Domain/CustomOrderDefaults.cs
@@ -1,0 +1,16 @@
+﻿namespace Nop.Core.CustomCode.Domain;
+
+/// <summary>
+/// Represents default values related to custom order data
+/// </summary>
+public static partial class CustomOrderDefaults
+{
+    #region Checkout
+
+    /// <summary>
+    /// Gets a name of generic attribute to store the value of 'GiftMessage'
+    /// </summary>
+    public static string GiftMessageGenericAttributeKey => "GiftMessage";
+
+    #endregion
+}
