+++ src/Presentation/Nop.Web/Areas/Admin/Factories/OrderModelFactory.cs
@@ -1160,72 +1160,72 @@ public partial class OrderModelFactory : IOrderModelFactory
         return model;
     }
 
-    /// <summary>
-    /// Prepare order model
-    /// </summary>
-    /// <param name="model">Order model</param>
-    /// <param name="order">Order</param>
-    /// <param name="excludeProperties">Whether to exclude populating of some properties of model</param>
-    /// <returns>
-    /// A task that represents the asynchronous operation
-    /// The task result contains the order model
-    /// </returns>
-    public virtual async Task<OrderModel> PrepareOrderModelAsync(OrderModel model, Order order, bool excludeProperties = false)
-    {
-        if (order != null)
-        {
-            //fill in model values from the entity
-            model ??= new OrderModel
-            {
-                Id = order.Id,
-                OrderStatusId = order.OrderStatusId,
-                VatNumber = order.VatNumber,
-                CheckoutAttributeInfo = order.CheckoutAttributeDescription
-            };
+    ///// <summary>
+    ///// Prepare order model
+    ///// </summary>
+    ///// <param name="model">Order model</param>
+    ///// <param name="order">Order</param>
+    ///// <param name="excludeProperties">Whether to exclude populating of some properties of model</param>
+    ///// <returns>
+    ///// A task that represents the asynchronous operation
+    ///// The task result contains the order model
+    ///// </returns>
+    //public virtual async Task<OrderModel> PrepareOrderModelAsync(OrderModel model, Order order, bool excludeProperties = false)
+    //{
+    //    if (order != null)
+    //    {
+    //        //fill in model values from the entity
+    //        model ??= new OrderModel
+    //        {
+    //            Id = order.Id,
+    //            OrderStatusId = order.OrderStatusId,
+    //            VatNumber = order.VatNumber,
+    //            CheckoutAttributeInfo = order.CheckoutAttributeDescription
+    //        };
 
-            var customer = await _customerService.GetCustomerByIdAsync(order.CustomerId);
+    //        var customer = await _customerService.GetCustomerByIdAsync(order.CustomerId);
 
-            model.OrderGuid = order.OrderGuid;
-            model.CustomOrderNumber = order.CustomOrderNumber;
-            model.CustomerIp = order.CustomerIp;
-            model.CustomerId = customer.Id;
-            model.OrderStatus = await _localizationService.GetLocalizedEnumAsync(order.OrderStatus);
-            model.StoreName = (await _storeService.GetStoreByIdAsync(order.StoreId))?.Name ?? "Deleted";
-            model.CustomerInfo = await _customerService.IsRegisteredAsync(customer) ? customer.Email : await _localizationService.GetResourceAsync("Admin.Customers.Guest");
-            model.CreatedOn = await _dateTimeHelper.ConvertToUserTimeAsync(order.CreatedOnUtc, DateTimeKind.Utc);
-            model.CustomValues = _paymentService.DeserializeCustomValues(order);
+    //        model.OrderGuid = order.OrderGuid;
+    //        model.CustomOrderNumber = order.CustomOrderNumber;
+    //        model.CustomerIp = order.CustomerIp;
+    //        model.CustomerId = customer.Id;
+    //        model.OrderStatus = await _localizationService.GetLocalizedEnumAsync(order.OrderStatus);
+    //        model.StoreName = (await _storeService.GetStoreByIdAsync(order.StoreId))?.Name ?? "Deleted";
+    //        model.CustomerInfo = await _customerService.IsRegisteredAsync(customer) ? customer.Email : await _localizationService.GetResourceAsync("Admin.Customers.Guest");
+    //        model.CreatedOn = await _dateTimeHelper.ConvertToUserTimeAsync(order.CreatedOnUtc, DateTimeKind.Utc);
+    //        model.CustomValues = _paymentService.DeserializeCustomValues(order);
 
-            var affiliate = await _affiliateService.GetAffiliateByIdAsync(order.AffiliateId);
-            if (affiliate != null)
-            {
-                model.AffiliateId = affiliate.Id;
-                model.AffiliateName = await _affiliateService.GetAffiliateFullNameAsync(affiliate);
-            }
+    //        var affiliate = await _affiliateService.GetAffiliateByIdAsync(order.AffiliateId);
+    //        if (affiliate != null)
+    //        {
+    //            model.AffiliateId = affiliate.Id;
+    //            model.AffiliateName = await _affiliateService.GetAffiliateFullNameAsync(affiliate);
+    //        }
 
-            //prepare order totals
-            await PrepareOrderModelTotalsAsync(model, order);
+    //        //prepare order totals
+    //        await PrepareOrderModelTotalsAsync(model, order);
 
-            //prepare order items
-            await PrepareOrderItemModelsAsync(model.Items, order);
-            model.HasDownloadableProducts = model.Items.Any(item => item.IsDownload);
+    //        //prepare order items
+    //        await PrepareOrderItemModelsAsync(model.Items, order);
+    //        model.HasDownloadableProducts = model.Items.Any(item => item.IsDownload);
 
-            //prepare payment info
-            await PrepareOrderModelPaymentInfoAsync(model, order);
+    //        //prepare payment info
+    //        await PrepareOrderModelPaymentInfoAsync(model, order);
 
-            //prepare shipping info
-            await PrepareOrderModelShippingInfoAsync(model, order);
+    //        //prepare shipping info
+    //        await PrepareOrderModelShippingInfoAsync(model, order);
 
-            //prepare nested search model
-            PrepareOrderShipmentSearchModel(model.OrderShipmentSearchModel, order);
-            PrepareOrderNoteSearchModel(model.OrderNoteSearchModel, order);
-        }
+    //        //prepare nested search model
+    //        PrepareOrderShipmentSearchModel(model.OrderShipmentSearchModel, order);
+    //        PrepareOrderNoteSearchModel(model.OrderNoteSearchModel, order);
+    //    }
 
-        model.IsLoggedInAsVendor = await _workContext.GetCurrentVendorAsync() != null;
-        model.AllowCustomersToSelectTaxDisplayType = _taxSettings.AllowCustomersToSelectTaxDisplayType;
-        model.TaxDisplayType = _taxSettings.TaxDisplayType;
+    //    model.IsLoggedInAsVendor = await _workContext.GetCurrentVendorAsync() != null;
+    //    model.AllowCustomersToSelectTaxDisplayType = _taxSettings.AllowCustomersToSelectTaxDisplayType;
+    //    model.TaxDisplayType = _taxSettings.TaxDisplayType;
 
-        return model;
-    }
+    //    return model;
+    //}
 
     /// <summary>
     /// Prepare upload license model
