+++ src/Presentation/Nop.Web/Controllers/CheckoutController.cs
@@ -2007,121 +2007,121 @@ public partial class CheckoutController : BasePublicController
         }
     }
 
-    [ValidateCaptcha]
-    [HttpPost]
-    public virtual async Task<IActionResult> OpcConfirmOrder(bool captchaValid)
-    {
-        try
-        {
-            var customer = await _workContext.GetCurrentCustomerAsync();
+    //[ValidateCaptcha]
+    //[HttpPost]
+    //public virtual async Task<IActionResult> OpcConfirmOrder(bool captchaValid)
+    //{
+    //    try
+    //    {
+    //        var customer = await _workContext.GetCurrentCustomerAsync();
 
-            var isCaptchaSettingEnabled = await _customerService.IsGuestAsync(customer) &&
-                                          _captchaSettings.Enabled && _captchaSettings.ShowOnCheckoutPageForGuests;
+    //        var isCaptchaSettingEnabled = await _customerService.IsGuestAsync(customer) &&
+    //                                      _captchaSettings.Enabled && _captchaSettings.ShowOnCheckoutPageForGuests;
 
-            var confirmOrderModel = new CheckoutConfirmModel()
-            {
-                DisplayCaptcha = isCaptchaSettingEnabled
-            };
+    //        var confirmOrderModel = new CheckoutConfirmModel()
+    //        {
+    //            DisplayCaptcha = isCaptchaSettingEnabled
+    //        };
 
-            //captcha validation for guest customers
-            if (!isCaptchaSettingEnabled || (isCaptchaSettingEnabled && captchaValid))
-            {
-                //validation
-                if (_orderSettings.CheckoutDisabled)
-                    throw new Exception(await _localizationService.GetResourceAsync("Checkout.Disabled"));
+    //        //captcha validation for guest customers
+    //        if (!isCaptchaSettingEnabled || (isCaptchaSettingEnabled && captchaValid))
+    //        {
+    //            //validation
+    //            if (_orderSettings.CheckoutDisabled)
+    //                throw new Exception(await _localizationService.GetResourceAsync("Checkout.Disabled"));
 
-                var store = await _storeContext.GetCurrentStoreAsync();
-                var cart = await _shoppingCartService.GetShoppingCartAsync(customer, ShoppingCartType.ShoppingCart, store.Id);
+    //            var store = await _storeContext.GetCurrentStoreAsync();
+    //            var cart = await _shoppingCartService.GetShoppingCartAsync(customer, ShoppingCartType.ShoppingCart, store.Id);
 
-                if (!cart.Any())
-                    throw new Exception("Your cart is empty");
+    //            if (!cart.Any())
+    //                throw new Exception("Your cart is empty");
 
-                if (!_orderSettings.OnePageCheckoutEnabled)
-                    throw new Exception("One page checkout is disabled");
+    //            if (!_orderSettings.OnePageCheckoutEnabled)
+    //                throw new Exception("One page checkout is disabled");
 
-                if (await _customerService.IsGuestAsync(customer) && !_orderSettings.AnonymousCheckoutAllowed)
-                    throw new Exception("Anonymous checkout is not allowed");
+    //            if (await _customerService.IsGuestAsync(customer) && !_orderSettings.AnonymousCheckoutAllowed)
+    //                throw new Exception("Anonymous checkout is not allowed");
 
-                //prevent 2 orders being placed within an X seconds time frame
-                if (!await IsMinimumOrderPlacementIntervalValidAsync(customer))
-                    throw new Exception(await _localizationService.GetResourceAsync("Checkout.MinOrderPlacementInterval"));
+    //            //prevent 2 orders being placed within an X seconds time frame
+    //            if (!await IsMinimumOrderPlacementIntervalValidAsync(customer))
+    //                throw new Exception(await _localizationService.GetResourceAsync("Checkout.MinOrderPlacementInterval"));
 
-                //place order
-                var processPaymentRequest = await HttpContext.Session.GetAsync<ProcessPaymentRequest>("OrderPaymentInfo");
-                if (processPaymentRequest == null)
-                {
-                    //Check whether payment workflow is required
-                    if (await _orderProcessingService.IsPaymentWorkflowRequiredAsync(cart))
-                    {
-                        throw new Exception("Payment information is not entered");
-                    }
+    //            //place order
+    //            var processPaymentRequest = await HttpContext.Session.GetAsync<ProcessPaymentRequest>("OrderPaymentInfo");
+    //            if (processPaymentRequest == null)
+    //            {
+    //                //Check whether payment workflow is required
+    //                if (await _orderProcessingService.IsPaymentWorkflowRequiredAsync(cart))
+    //                {
+    //                    throw new Exception("Payment information is not entered");
+    //                }
 
-                    processPaymentRequest = new ProcessPaymentRequest();
-                }
-                await _paymentService.GenerateOrderGuidAsync(processPaymentRequest);
-                processPaymentRequest.StoreId = store.Id;
-                processPaymentRequest.CustomerId = customer.Id;
-                processPaymentRequest.PaymentMethodSystemName = await _genericAttributeService.GetAttributeAsync<string>(customer,
-                    NopCustomerDefaults.SelectedPaymentMethodAttribute, store.Id);
-                await HttpContext.Session.SetAsync("OrderPaymentInfo", processPaymentRequest);
-                var placeOrderResult = await _orderProcessingService.PlaceOrderAsync(processPaymentRequest);
-                if (placeOrderResult.Success)
-                {
-                    await HttpContext.Session.SetAsync<ProcessPaymentRequest>("OrderPaymentInfo", null);
-                    var postProcessPaymentRequest = new PostProcessPaymentRequest
-                    {
-                        Order = placeOrderResult.PlacedOrder
-                    };
+    //                processPaymentRequest = new ProcessPaymentRequest();
+    //            }
+    //            await _paymentService.GenerateOrderGuidAsync(processPaymentRequest);
+    //            processPaymentRequest.StoreId = store.Id;
+    //            processPaymentRequest.CustomerId = customer.Id;
+    //            processPaymentRequest.PaymentMethodSystemName = await _genericAttributeService.GetAttributeAsync<string>(customer,
+    //                NopCustomerDefaults.SelectedPaymentMethodAttribute, store.Id);
+    //            await HttpContext.Session.SetAsync("OrderPaymentInfo", processPaymentRequest);
+    //            var placeOrderResult = await _orderProcessingService.PlaceOrderAsync(processPaymentRequest);
+    //            if (placeOrderResult.Success)
+    //            {
+    //                await HttpContext.Session.SetAsync<ProcessPaymentRequest>("OrderPaymentInfo", null);
+    //                var postProcessPaymentRequest = new PostProcessPaymentRequest
+    //                {
+    //                    Order = placeOrderResult.PlacedOrder
+    //                };
 
-                    var paymentMethod = await _paymentPluginManager
-                        .LoadPluginBySystemNameAsync(placeOrderResult.PlacedOrder.PaymentMethodSystemName, customer, store.Id);
-                    if (paymentMethod == null)
-                        //payment method could be null if order total is 0
-                        //success
-                        return Json(new { success = 1 });
+    //                var paymentMethod = await _paymentPluginManager
+    //                    .LoadPluginBySystemNameAsync(placeOrderResult.PlacedOrder.PaymentMethodSystemName, customer, store.Id);
+    //                if (paymentMethod == null)
+    //                    //payment method could be null if order total is 0
+    //                    //success
+    //                    return Json(new { success = 1 });
 
-                    if (paymentMethod.PaymentMethodType == PaymentMethodType.Redirection)
-                    {
-                        //Redirection will not work because it's AJAX request.
-                        //That's why we don't process it here (we redirect a user to another page where he'll be redirected)
+    //                if (paymentMethod.PaymentMethodType == PaymentMethodType.Redirection)
+    //                {
+    //                    //Redirection will not work because it's AJAX request.
+    //                    //That's why we don't process it here (we redirect a user to another page where he'll be redirected)
 
-                        //redirect
-                        return Json(new
-                        {
-                            redirect = $"{_webHelper.GetStoreLocation()}checkout/OpcCompleteRedirectionPayment"
-                        });
-                    }
+    //                    //redirect
+    //                    return Json(new
+    //                    {
+    //                        redirect = $"{_webHelper.GetStoreLocation()}checkout/OpcCompleteRedirectionPayment"
+    //                    });
+    //                }
 
-                    await _paymentService.PostProcessPaymentAsync(postProcessPaymentRequest);
-                    //success
-                    return Json(new { success = 1 });
-                }
+    //                await _paymentService.PostProcessPaymentAsync(postProcessPaymentRequest);
+    //                //success
+    //                return Json(new { success = 1 });
+    //            }
 
-                //error
-                foreach (var error in placeOrderResult.Errors)
-                    confirmOrderModel.Warnings.Add(error);
-            }
-            else
-            {
-                confirmOrderModel.Warnings.Add(await _localizationService.GetResourceAsync("Common.WrongCaptchaMessage"));
-            }
+    //            //error
+    //            foreach (var error in placeOrderResult.Errors)
+    //                confirmOrderModel.Warnings.Add(error);
+    //        }
+    //        else
+    //        {
+    //            confirmOrderModel.Warnings.Add(await _localizationService.GetResourceAsync("Common.WrongCaptchaMessage"));
+    //        }
 
-            return Json(new
-            {
-                update_section = new UpdateSectionJsonModel
-                {
-                    name = "confirm-order",
-                    html = await RenderPartialViewToStringAsync("OpcConfirmOrder", confirmOrderModel)
-                },
-                goto_section = "confirm_order"
-            });
-        }
-        catch (Exception exc)
-        {
-            await _logger.WarningAsync(exc.Message, exc, await _workContext.GetCurrentCustomerAsync());
-            return Json(new { error = 1, message = exc.Message });
-        }
-    }
+    //        return Json(new
+    //        {
+    //            update_section = new UpdateSectionJsonModel
+    //            {
+    //                name = "confirm-order",
+    //                html = await RenderPartialViewToStringAsync("OpcConfirmOrder", confirmOrderModel)
+    //            },
+    //            goto_section = "confirm_order"
+    //        });
+    //    }
+    //    catch (Exception exc)
+    //    {
+    //        await _logger.WarningAsync(exc.Message, exc, await _workContext.GetCurrentCustomerAsync());
+    //        return Json(new { error = 1, message = exc.Message });
+    //    }
+    //}
 
     public virtual async Task<IActionResult> OpcCompleteRedirectionPayment()
     {
