+++ src/Presentation/Nop.Web/CustomCode/Controllers/CheckoutController.cs
@@ -1,0 +1,145 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Nop.Core.CustomCode.Domain;
+using Nop.Core.Domain.Customers;
+using Nop.Core.Domain.Orders;
+using Nop.Core.Http.Extensions;
+using Nop.Services.Payments;
+using Nop.Web.Framework.Mvc.Filters;
+using Nop.Web.Models.Checkout;
+
+namespace Nop.Web.Controllers;
+
+public partial class CheckoutController
+{
+    #region Methods (one page checkout)
+
+    [ValidateCaptcha]
+    [HttpPost]
+    public virtual async Task<IActionResult> OpcConfirmOrder(bool captchaValid, IFormCollection form)
+    {
+        try
+        {
+            var customer = await _workContext.GetCurrentCustomerAsync();
+
+            var isCaptchaSettingEnabled = await _customerService.IsGuestAsync(customer) &&
+                                          _captchaSettings.Enabled && _captchaSettings.ShowOnCheckoutPageForGuests;
+
+            var confirmOrderModel = new CheckoutConfirmModel()
+            {
+                DisplayCaptcha = isCaptchaSettingEnabled
+            };
+
+            //captcha validation for guest customers
+            if (!isCaptchaSettingEnabled || (isCaptchaSettingEnabled && captchaValid))
+            {
+                //validation
+                if (_orderSettings.CheckoutDisabled)
+                    throw new Exception(await _localizationService.GetResourceAsync("Checkout.Disabled"));
+
+                var store = await _storeContext.GetCurrentStoreAsync();
+                var cart = await _shoppingCartService.GetShoppingCartAsync(customer, ShoppingCartType.ShoppingCart, store.Id);
+
+                if (!cart.Any())
+                    throw new Exception("Your cart is empty");
+
+                if (!_orderSettings.OnePageCheckoutEnabled)
+                    throw new Exception("One page checkout is disabled");
+
+                if (await _customerService.IsGuestAsync(customer) && !_orderSettings.AnonymousCheckoutAllowed)
+                    throw new Exception("Anonymous checkout is not allowed");
+
+                //prevent 2 orders being placed within an X seconds time frame
+                if (!await IsMinimumOrderPlacementIntervalValidAsync(customer))
+                    throw new Exception(await _localizationService.GetResourceAsync("Checkout.MinOrderPlacementInterval"));
+
+                //place order
+                var processPaymentRequest = await HttpContext.Session.GetAsync<ProcessPaymentRequest>("OrderPaymentInfo");
+                if (processPaymentRequest == null)
+                {
+                    //Check whether payment workflow is required
+                    if (await _orderProcessingService.IsPaymentWorkflowRequiredAsync(cart))
+                    {
+                        throw new Exception("Payment information is not entered");
+                    }
+
+                    processPaymentRequest = new ProcessPaymentRequest();
+                }
+                await _paymentService.GenerateOrderGuidAsync(processPaymentRequest);
+                processPaymentRequest.StoreId = store.Id;
+                processPaymentRequest.CustomerId = customer.Id;
+                processPaymentRequest.PaymentMethodSystemName = await _genericAttributeService.GetAttributeAsync<string>(customer,
+                    NopCustomerDefaults.SelectedPaymentMethodAttribute, store.Id);
+                await HttpContext.Session.SetAsync("OrderPaymentInfo", processPaymentRequest);
+                var placeOrderResult = await _orderProcessingService.PlaceOrderAsync(processPaymentRequest);
+                if (placeOrderResult.Success)
+                {
+                    await HttpContext.Session.SetAsync<ProcessPaymentRequest>("OrderPaymentInfo", null);
+                    var postProcessPaymentRequest = new PostProcessPaymentRequest
+                    {
+                        Order = placeOrderResult.PlacedOrder
+                    };
+
+                    //++Custom code
+
+                    //save gift message in order's generic attribute
+                    if (form.TryGetValue(CustomOrderDefaults.GiftMessageGenericAttributeKey, out var message) && !string.IsNullOrEmpty(message))
+                    {
+                        await _genericAttributeService.SaveAttributeAsync(placeOrderResult.PlacedOrder,
+                            CustomOrderDefaults.GiftMessageGenericAttributeKey, message.ToString(), store.Id);
+                    }
+
+                    //--Custom code
+
+                    var paymentMethod = await _paymentPluginManager
+                        .LoadPluginBySystemNameAsync(placeOrderResult.PlacedOrder.PaymentMethodSystemName, customer, store.Id);
+                    if (paymentMethod == null)
+                        //payment method could be null if order total is 0
+                        //success
+                        return Json(new { success = 1 });
+
+                    if (paymentMethod.PaymentMethodType == PaymentMethodType.Redirection)
+                    {
+                        //Redirection will not work because it's AJAX request.
+                        //That's why we don't process it here (we redirect a user to another page where he'll be redirected)
+
+                        //redirect
+                        return Json(new
+                        {
+                            redirect = $"{_webHelper.GetStoreLocation()}checkout/OpcCompleteRedirectionPayment"
+                        });
+                    }
+
+                    await _paymentService.PostProcessPaymentAsync(postProcessPaymentRequest);
+                    //success
+                    return Json(new { success = 1 });
+                }
+
+                //error
+                foreach (var error in placeOrderResult.Errors)
+                    confirmOrderModel.Warnings.Add(error);
+            }
+            else
+            {
+                confirmOrderModel.Warnings.Add(await _localizationService.GetResourceAsync("Common.WrongCaptchaMessage"));
+            }
+
+            return Json(new
+            {
+                update_section = new UpdateSectionJsonModel
+                {
+                    name = "confirm-order",
+                    html = await RenderPartialViewToStringAsync("OpcConfirmOrder", confirmOrderModel)
+                },
+                goto_section = "confirm_order"
+            });
+        }
+        catch (Exception exc)
+        {
+            await _logger.WarningAsync(exc.Message, exc, await _workContext.GetCurrentCustomerAsync());
+            return Json(new { error = 1, message = exc.Message });
+        }
+    }
+
+
+    #endregion
+}
