+++ src/Presentation/Nop.Web/CustomCode/Models/Order/OrderDetailsModel.cs
@@ -1,0 +1,6 @@
+﻿namespace Nop.Web.Models.Order;
+
+public partial record OrderDetailsModel 
+{
+    public string GiftMessage { get; set; }
+}
